fix norm

+++ philo/safe.c
@@ -12,49 +12,49 @@
 
 #include"philo.h"
 
-int should_stop(t_simulation *sim)
+int	should_stop(t_simulation *sim)
 {
-  int value;
+	int	value;
 
-  pthread_mutex_lock(&sim->stop_lock);
-  value = sim->stop;
-  pthread_mutex_unlock(&sim->stop_lock);
-  return (value);
+	pthread_mutex_lock(&sim->stop_lock);
+	value = sim->stop;
+	pthread_mutex_unlock(&sim->stop_lock);
+	return (value);
 }
 
-void set_stop(t_simulation *sim)
+void	set_stop(t_simulation *sim)
 {
-  pthread_mutex_lock(&sim->stop_lock);
-  sim->stop = 1;
-  pthread_mutex_unlock(&sim->stop_lock);
+	pthread_mutex_lock(&sim->stop_lock);
+	sim->stop = 1;
+	pthread_mutex_unlock(&sim->stop_lock);
 }
 
-long long get_last_meal(t_philo *philo)
+size_t	get_last_meal(t_philo *philo)
 {
-  long long time;
+	size_t	time;
 
-  pthread_mutex_lock(&philo->simul->meal_lock);
-  time = philo->last_meal;
-  pthread_mutex_unlock(&philo->simul->meal_lock);
-  return (time);
+	pthread_mutex_lock(&philo->simul->meal_lock);
+	time = philo->last_meal;
+	pthread_mutex_unlock(&philo->simul->meal_lock);
+	return (time);
 }
 
-void set_last_meal(t_philo *philo, long long time)
+void	set_last_meal(t_philo *philo, long long time)
 {
-  pthread_mutex_lock(&philo->simul->meal_lock);
-  philo->last_meal = time;
-  pthread_mutex_unlock(&philo->simul->meal_lock);
+	pthread_mutex_lock(&philo->simul->meal_lock);
+	philo->last_meal = time;
+	pthread_mutex_unlock(&philo->simul->meal_lock);
 }
 
-void safe_print(t_simulation *sim, int philo_id, char *message)
+void	safe_print(t_simulation *sim, int philo_id, char *message)
 {
-  size_t timestamp;
-    
-  pthread_mutex_lock(&sim->print_lock);
-  if (!should_stop(sim))
-  {
-    timestamp = get_time() - sim->start_time;
-    printf("%zu %d %s\n", timestamp, philo_id, message);
-  }
-  pthread_mutex_unlock(&sim->print_lock);
+	size_t	timestamp;
+
+	pthread_mutex_lock(&sim->print_lock);
+	if (!should_stop(sim))
+	{
+		timestamp = get_time() - sim->start_time;
+		printf("%zu %d %s\n", timestamp, philo_id, message);
+	}
+	pthread_mutex_unlock(&sim->print_lock);
 }
